Resolve unnecessary_cast clippy lint
    error: casting to the same type is unnecessary (`u64` -> `u64`)
      --> src/alloc.rs:89:28
       |
    89 |         total: ByteSize::b(ALLOC.total.load(Ordering::Relaxed) as u64),
       |                            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `ALLOC.total.load(Ordering::Relaxed)`
       |
       = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
       = note: `-D clippy::unnecessary-cast` implied by `-D clippy::all`

    error: casting to the same type is unnecessary (`u64` -> `u64`)
      --> src/alloc.rs:90:27
       |
    90 |         peak: ByteSize::b(ALLOC.peak.load(Ordering::Relaxed) as u64),
       |                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `ALLOC.peak.load(Ordering::Relaxed)`
       |
       = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast

+++ src/alloc.rs
@@ -86,8 +86,8 @@ pub(crate) struct AllocStat {
 pub(crate) fn stat() -> AllocStat {
     AllocStat {
         count: ALLOC.count.load(Ordering::Relaxed),
-        total: ByteSize::b(ALLOC.total.load(Ordering::Relaxed) as u64),
-        peak: ByteSize::b(ALLOC.peak.load(Ordering::Relaxed) as u64),
+        total: ByteSize::b(ALLOC.total.load(Ordering::Relaxed)),
+        peak: ByteSize::b(ALLOC.peak.load(Ordering::Relaxed)),
     }
 }
